requirements.txt
requests==2.31.0
python-dotenv==1.0.0
Flask==3.0.1
Flask-WTF==1.2.1
Werkzeug==3.0.1
flask-cors==4.0.0
